% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{package}
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline Label Geom for Annotating Earthquakes in earthquake_R}
\usage{
geom_timeline_label(mapping = NULL, data = NULL,
                    stat = "identity", position = "identity",
                    ..., na.rm = TRUE, n_max = NULL,
                    show.legend = NA, inherit.aes = TRUE)
}
\description{
\code{geom_timeline_label()} is a ggplot2 geom provided by the \strong{earthquake_R} package.
It is designed to add text annotations to a timeline plot by labeling the largest earthquakes.
}
\details{
This geom leverages the \code{GeomTimelineLabel} ggproto object to draw vertical lines from each earthquake point and attach text labels to annotate the most significant events on the timeline. The \code{n_max} parameter allows you to restrict the number of annotations to only the top \emph{n} earthquakes (based on a specified metric such as magnitude or number of deaths). It is typically used alongside \code{geom_timeline()} to enhance the visual presentation of earthquake data.
}
\seealso{
\code{\link{geom_timeline}}, \code{\link{theme_timeline}}
}
\examples{
library(ggplot2)
library(dplyr)
library(stringr)

# Load and clean example earthquake data
raw_data <- readr::read_delim(
  system.file("extdata/earthquakes.tsv.gz", package = "earthquake_R"),
  delim = "\t"
)
clean_data <- eq_clean_data(raw_data)

# Create a timeline plot with annotations for the largest earthquakes
g <- clean_data %>%
  mutate_at(vars(DEATHS, EQ_PRIMARY), as.numeric) %>%
  filter(str_detect(tolower(COUNTRY), "china|usa$|pakistan")) %>%
  filter(DATE >= "2000-01-01") %>%
  ggplot(aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, fill = DEATHS))
g + geom_timeline() +
    geom_timeline_label(aes(label = LOCATION_NAME, n_max_var = DEATHS), n_max = 5)
}
\keyword{plot}
