% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAA_data}
\alias{NOAA_data}
\title{NOAA Significant Earthquake Database}
\format{
A data frame with 6219 observations and 39 variables.
}
\source{
NOAA National Centers for Environmental Information, National Geophysical Data Center / World Data Service (NGDC/WDS):\\
NCEI/WDS Global Significant Earthquake Database. NOAA National Centers for Environmental Information.\\
doi:10.7289/V5TD9V7K\\
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
NOAA_data
}
\description{
This dataset was obtained from the U.S. National Oceanic and Atmospheric Administration (NOAA) on significant earthquakes around the world. It contains information about 5,933 earthquakes over an approximately 4,000-year time span. The data is utilized in the \strong{earthquake_R} package to demonstrate data cleaning, visualization, and analysis of historical earthquake events.
}
\keyword{datasets}
